Dockerized local development environment
