+++ README.md
@@ -5,8 +5,20 @@
 - `BOT_TOKEN`: the token found in the bot user account.
 - `NOTIF_CHANNEL_ID`: the ID of the channel the bot will send system notifications to.
 
+## Prerequisites
+
+- [Yarn](https://classic.yarnpkg.com/en/docs/install)
+- [Docker](https://docs.docker.com/get-docker/) (tested up to v20.10.6)
+
 ## Running the bot locally
 
-1. Run `yarn` to install dependencies.
-1. Add the required environment variables in a `.env` file in the root directory.
-1. Run `yarn dev` to start the bot locally.
+1. Build docker image: `yarn image:build`
+1. Start container in detached mode: `yarn start`
+1. View and follow console output: `yarn logs`
+
+## Other usage
+
+- Stop the container: `yarn stop`
+- Stop and remove the container: `yarn clean`
+- Restart the container: `yarn restart`
+- Fresh build and restart: `yarn image:build && yarn clean && yarn start`
