+++ package.json
@@ -4,10 +4,16 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "concurrently \"yarn run-dev\" \"yarn watch\"",
-    "run-dev": "cross-env NODE_ENV=dev nodemon index.ts",
-    "watch": "tsc --watch",
-    "build": "tsc",
+    "local:run": "concurrently \"yarn node:watch\" \"yarn ts:watch\"",
+    "node:watch": "NODE_ENV=dev nodemon index.ts",
+    "ts:watch": "tsc --watch",
+    "ts:build": "tsc",
+    "image:build": "docker build . -t codey",
+    "clean": "docker compose down",
+    "stop": "docker compose stop",
+    "start": "docker compose up -d",
+    "restart": "docker compose restart",
+    "logs": "docker logs -f codey-bot",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
