+++ src/components/db.ts
@@ -1,22 +1,22 @@
-import sqlite3 = require('sqlite3')
-import { open, Database } from 'sqlite'
-import Discord from 'discord.js'
+import sqlite3 = require('sqlite3');
+import { open, Database } from 'sqlite';
+import Discord from 'discord.js';
 
-let db : Database | null = null;
+let db: Database | null = null;
 
-export async function openDB () {
-  if(db == null){
+export async function openDB() {
+  if (db == null) {
     db = await open({
-      filename: './db/bot.db',
+      filename: 'db/bot.db',
       driver: sqlite3.Database
-    })
-    await db.run('CREATE TABLE IF NOT EXISTS saved_data (msg_id INTEGER PRIMARY KEY,data TEXT NOT NULL);')
+    });
+    await db.run('CREATE TABLE IF NOT EXISTS saved_data (msg_id INTEGER PRIMARY KEY,data TEXT NOT NULL);');
   }
   return db;
 }
 
-export async function testDb(message: Discord.Message, command: string, args: string[]){
-  switch(command){
+export async function testDb(message: Discord.Message, command: string, args: string[]) {
+  switch (command) {
     case 'save':
       if (args.length < 1) {
         await message.channel.send('no args');
@@ -35,7 +35,7 @@ export async function testDb(message: Discord.Message, command: string, args: st
           .setTitle('Database Dump')
           .setURL('https://www.youtube.com/watch?v=dQw4w9WgXcQ');
         const res = await db.all('SELECT * FROM saved_data');
-        for(const rows of res){
+        for (const rows of res) {
           console.log(rows['msg_id'], rows['data']);
           outEmbed = outEmbed.addField(rows['msg_id'], rows['data'], true);
           console.log(outEmbed);
@@ -65,4 +65,4 @@ export async function testDb(message: Discord.Message, command: string, args: st
   }
 }
 
-console.log('connected to db')
+console.log('connected to db');
